Move (m/o)extra tables into their .c files (from .h)
They are only used in their individual .c's. Other files do not and should not need to know what the tables look like.
