+++ include/oextra.h
@@ -22,15 +22,4 @@ union oextra {
 	};
 };
 
-struct ox_table {
-	int indexnum;
-	int s_size;
-};
-
-static struct ox_table ox_list[] = {
-	{OX_ENAM, -1},	/* variable; actual size is stored in structure. 1st item is an int containing size */
-	{OX_EMON, -1},	/* variable; actual size is stored in structure. 1st item is an int containing size */
-	{OX_EMID, sizeof(int)}
-};
-
 #endif /* OEXTRA_H */
